Remove knowledge of all goals in handlers
