+++ src/handlers/events/delivery/goals/httpServiceGoals.ts
@@ -66,12 +66,12 @@ export const StagingVerifiedGoal = new Goal({
     completedDescription: "Verified endpoint in Test",
 });
 
-export const ProductionDeploymentGoal = new Goal({
+export const ProductionDeploymentGoal = new GoalWithPrecondition({
     environment: ProductionEnvironment,
     orderedName: "3-prod-deploy",
     displayName: "deploy to Prod",
     completedDescription: "Deployed to Prod",
-});
+}, StagingDeploymentGoal);
 
 export const ProductionEndpointGoal = new Goal({
     environment: ProductionEnvironment,
