+++ src/software-delivery-machine/blueprint/deploy/k8sDeploy.ts
@@ -1,23 +1,24 @@
 import { NoticeK8sTestDeployCompletionOnStatus } from "../../../handlers/events/delivery/deploy/k8s/NoticeK8sDeployCompletion";
 import { NoticeK8sProdDeployCompletionOnStatus } from "../../../handlers/events/delivery/deploy/k8s/NoticeK8sProdDeployCompletion";
-import { RequestK8sDeployOnSuccessStatus } from "../../../handlers/events/delivery/deploy/k8s/RequestDeployOnSuccessStatus";
-import { RequestK8sDeployOnSuccessStatus1 } from "../../../handlers/events/delivery/deploy/k8s/RequestDeployOnSuccessStatus1";
+import { RequestK8sDeployOnSuccessStatus } from "../../../handlers/events/delivery/deploy/k8s/RequestK8sDeployOnSuccessStatus";
+import { RequestK8sDeployOnSuccessStatus1 } from "../../../handlers/events/delivery/deploy/k8s/RequestK8sDeployOnSuccessStatus1";
 import {
-    ContextToPlannedGoal, HttpServiceGoals, ProductionDeploymentContext, ProductionEndpointContext, StagingDeploymentContext,
+    ContextToPlannedGoal,
+    ProductionDeploymentContext,
+    ProductionEndpointContext,
+    StagingDeploymentContext,
     StagingEndpointContext,
 } from "../../../handlers/events/delivery/goals/httpServiceGoals";
 import { K8sProductionDomain, K8sTestingDomain } from "./describeRunningServices";
 
 export const K8sStagingDeployOnSuccessStatus = () =>
     new RequestK8sDeployOnSuccessStatus(
-        HttpServiceGoals,
         ContextToPlannedGoal[StagingDeploymentContext],
         K8sTestingDomain);
 
 export const K8sProductionDeployOnSuccessStatus = () =>
     // TODO replace this evil hack of the duplicate class
     new RequestK8sDeployOnSuccessStatus1(
-        HttpServiceGoals,
         ContextToPlannedGoal[ProductionDeploymentContext],
         K8sProductionDomain);
